+++ lib/src/model/branch_info.dart
@@ -24,22 +24,12 @@ class BranchInfo {
   // jsonName: 'shedule_description'
   final String scheduleDescription;
 
-  BranchInfo(
-      this.number,
-      this.address,
-      this.locality,
-      this.type,
-      this.format,
-      this.deliveryBranchId,
-      this.maxWeight,
-      this.latitude,
-      this.longitude,
-      this.description,
-      this.scheduleDescription);
+  BranchInfo(this.number, this.address, this.locality, this.type, this.format, this.deliveryBranchId, this.maxWeight,
+      this.latitude, this.longitude, this.description, this.scheduleDescription);
 
   @override
   String toString() {
-    return 'BranchInfo { '
+    return 'BranchInfo {'
         'number: $number, '
         'address: $address, '
         'locality: $locality, '