+++ lib/src/model/converter/branch_type.converter.dart
@@ -4,13 +4,7 @@ import 'package:djustin/src/model/converter/json_converter.dart';
 
 class BranchTypeConverter extends JsonConverter<BranchType> {
   @override
-  BranchType fromJson(Map<String, dynamic> json) {
-    BranchType branchType;
-    if (json != null) {
-      branchType = BranchType(
-          BranchFormatExtension.fromString(json['short_name']),
-          json['description']);
-    }
-    return branchType;
+  BranchType fromNotBlankJson(Map<String, dynamic> json) {
+    return BranchType(BranchFormatExtension.fromString(json['short_name']), json['description']);
   }
 }