+++ lib/src/model/converter/json_converter.dart
@@ -1,7 +1,11 @@
 import 'dart:convert' as json;
 
 abstract class JsonConverter<T> {
-  T fromJson(Map<String, dynamic> json);
+  T fromNotBlankJson(Map<String, dynamic> json);
+
+  T fromJson(Map<String, dynamic> json) {
+    return (json == null || json.isEmpty) ? null : fromNotBlankJson(json);
+  }
 
   T fromJsonString(String jsonString) {
     return fromJson(json.jsonDecode(jsonString));