+++ lib/src/model/converter/locality.converter.dart
@@ -4,23 +4,10 @@ import 'package:djustin/src/model/locality.dart';
 
 class LocalityConverter extends JsonConverter<Locality> {
   @override
-  Locality fromJson(Map<String, dynamic> json) {
-    Locality locality;
-    if (json != null && json.isNotEmpty) {
-      var title = <Language, String>{
-        Language.UA: json['title_ua'],
-        Language.EN: json['title_en'],
-        Language.RU: json['title_ru']
-      };
-      var parentTitle = <Language, String>{
-        Language.UA: json['parent_title_ua'],
-        Language.EN: json['parent_title_en'],
-        Language.RU: json['parent_title_ru']
-      };
+  Locality fromNotBlankJson(Map<String, dynamic> json) {
+    var title = Locale.parseLocalizedText('title_', json);
+    var parentTitle = Locale.parseLocalizedText('parent_title_', json);
 
-      locality = Locality(json['uuid'], json['SCOATOU'], json['parent_uuid'],
-          title, parentTitle);
-    }
-    return locality;
+    return Locality(json['uuid'], json['SCOATOU'], json['parent_uuid'], title, parentTitle);
   }
 }