+++ lib/src/model/locality.dart
@@ -15,17 +15,16 @@ class Locality {
   // jsonName: parent_title_ua, parent_title_en, parent_title_ru
   final Map<Language, String> parentTitle;
 
-  Locality(
-      this.uuid, this.scoatou, this.parentUuid, this.title, this.parentTitle);
+  Locality(this.uuid, this.scoatou, this.parentUuid, this.title, this.parentTitle);
 
   @override
   String toString() {
-    return 'Locality { '
+    return 'Locality {'
         'uuid: $uuid, '
         'scoatou: $scoatou, '
         'parentUuid: $parentUuid, '
         'title: $title, '
-        'parentTitle: $parentTitle '
+        'parentTitle: $parentTitle'
         '}';
   }
 }