+++ lib/src/model/service_info.dart
@@ -19,19 +19,19 @@ class ServiceInfo {
   // jsonName: send_service
   final bool hasSendService;
 
-  ServiceInfo(this.id, this.name, this.description, this.alias, this.hasSelfService,
-      this.hasCategoryService, this.hasSendService);
+  ServiceInfo(this.id, this.name, this.description, this.alias, this.hasSelfService, this.hasCategoryService,
+      this.hasSendService);
 
   @override
   String toString() {
-    return 'ServiceInfo { '
+    return 'ServiceInfo {'
         'id: $id, '
         'name: $name, '
         'description: $description, '
         'alias: $alias, '
         'hasSelfService: $hasSelfService, '
         'hasCategoryService: $hasCategoryService, '
-        'hasSendService: $hasSendService '
+        'hasSendService: $hasSendService'
         '}';
   }
 }