+++ lib/src/model/tracking.dart
@@ -8,18 +8,18 @@ class Tracking {
   final String departmentNumber;
   final String departmentAddress;
 
-  Tracking(this.orderNumber, this.orderDescription, this.dateTime, this.status,
-      this.departmentNumber, this.departmentAddress);
+  Tracking(this.orderNumber, this.orderDescription, this.dateTime, this.status, this.departmentNumber,
+      this.departmentAddress);
 
   @override
   String toString() {
-    return 'Tracking { '
+    return 'Tracking {'
         'orderNumber: $orderNumber, '
         'orderDescription: $orderDescription, '
         'dateTime: $dateTime, '
         'status: $status, '
         'departmentNumber: $departmentNumber, '
-        'departmentAddress: $departmentAddress '
+        'departmentAddress: $departmentAddress'
         '}';
   }
 }