+++ lib/src/service/justin_endpoints.dart
@@ -1,0 +1,10 @@
+library justin_endpoints;
+
+const String justinBase = 'http://openapi.justin.ua';
+const String justinBranches = '/branches';
+const String justinBranchTypes = '/branch_types';
+const String justinBranchLocator = '/branches_locator';
+const String justinTracking = '/tracking';
+const String justinTrackingHistory = '/tracking_history';
+const String justinLocalities = '/localities';
+const String justinServicesInfo = '/services';